+++ src/uio/utility/math/__init__.py
@@ -1,0 +1,3 @@
+"""
+.. include:: ../../../../documentation/utility/math/index.md
+"""